+++ src/Rik.StatusPage/Providers/MsSqlDatabaseStatusProvider.cs
@@ -1,10 +1,13 @@
 ﻿using System.Collections.Generic;
+using System.Reflection.Emit;
 using Rik.StatusPage.Configuration;
 
 namespace Rik.StatusPage.Providers
 {
     public class MsSqlDatabaseStatusProvider : DatabaseStatusProvider
     {
+        private static readonly GetDatabaseUriDelegate getDatabaseUri = CreateGetDatabaseUri();
+
         protected override IEnumerable<string> ConnectionTypeNames => new []
         {
             "System.Data.SqlClient.SqlConnection, System.Data"
@@ -16,5 +19,35 @@ namespace Rik.StatusPage.Providers
         public MsSqlDatabaseStatusProvider(StatusProviderConfigurationElement configuration)
             : base(configuration)
         { }
+
+        protected override string GetUri()
+        {
+            return getDatabaseUri(configuration.ConnectionString);
+        }
+
+        private static GetDatabaseUriDelegate CreateGetDatabaseUri()
+        {
+            var method = new DynamicMethod("MsSqlDatabaseStatusProvider_getDatabaseUri", typeof(string), new [] { typeof(string) });
+            var il = method.GetILGenerator();
+
+            var builderType = FindTypeOrFailWith("System.Data.SqlClient.SqlConnectionStringBuilder, System.Data", "Could not load type 'SqlConnectionStringBuilder'.");
+
+            // var builder = new SqlConnectionStringBuilder(connectionString);
+            var builder = il.DeclareLocal(builderType);
+            il.Emit(OpCodes.Ldarg_0);
+            il.Emit(OpCodes.Newobj, builderType.GetConstructor(new [] { typeof(string) }));
+            il.Emit(OpCodes.Stloc, builder);
+
+            // return string.Format("{0}@{1}", builder.InitialCatalog, builder.DataSource);
+            il.Emit(OpCodes.Ldstr, "{0}@{1}");
+            il.Emit(OpCodes.Ldloc, builder);
+            il.Emit(OpCodes.Callvirt, builderType.GetProperty("InitialCatalog").GetGetMethod());
+            il.Emit(OpCodes.Ldloc, builder);
+            il.Emit(OpCodes.Callvirt, builderType.GetProperty("DataSource").GetGetMethod());
+            il.Emit(OpCodes.Call, typeof(string).GetMethod("Format", new [] { typeof(string), typeof(object), typeof(object) }));
+            il.Emit(OpCodes.Ret);
+
+            return (GetDatabaseUriDelegate)method.CreateDelegate(typeof(GetDatabaseUriDelegate));
+        }
     }
 }